+++ Expense Tracker v2/ExpensesForm.cs
@@ -36,7 +36,7 @@ namespace Expense_Tracker_v2
         public void displayExpenseData()
         {
             ExpensesData eData = new ExpensesData();
-            List<ExpensesData> listData = eData.expensesListData();
+            List<ExpensesData> listData = eData.expensesListData(Loginn.UserId);
 
             bunifuDataGridView1.DataSource = listData;
         }
@@ -47,12 +47,13 @@ namespace Expense_Tracker_v2
             {
                 connection.Open();
 
-                string selectData = "SELECT category FROM categories WHERE type = @type AND status = @status";
+                string selectData = "SELECT category FROM categories WHERE type = @type AND status = @status AND UserId = @userId";
 
                 using (SqlCommand cmd = new SqlCommand(selectData, connection))
                 {
                     cmd.Parameters.AddWithValue("@type", "Expenses");
                     cmd.Parameters.AddWithValue("@status", "Active");
+                    cmd.Parameters.AddWithValue("@userId", Loginn.UserId);
 
                     expense_category.Items.Clear();
 
@@ -80,8 +81,8 @@ namespace Expense_Tracker_v2
                 {
                     connection.Open();
 
-                    string insertData = "INSERT INTO expenses (category, item, expense, description, date_expense, date_insert)" +
-                        "VALUES(@cat, @item, @expense, @desc, @date_ex, @date)";
+                    string insertData = "INSERT INTO expenses (category, item, expense, description, date_expense, date_insert, UserId)" +
+                        "VALUES(@cat, @item, @expense, @desc, @date_ex, @date, @userId)";
 
                     using (SqlCommand cmd = new SqlCommand(insertData, connection))
                     {
@@ -91,15 +92,12 @@ namespace Expense_Tracker_v2
                         cmd.Parameters.AddWithValue("@desc", expense_description.Text);
                         cmd.Parameters.AddWithValue("@date_ex", expense_date.Value);
                         cmd.Parameters.AddWithValue("@date", DateTime.Now.Date);
-
-                        
+                        cmd.Parameters.AddWithValue("@userId", Loginn.UserId);
 
                         cmd.ExecuteNonQuery();
                         clearFields();
 
                         MessageBox.Show("Expense added successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
-
                     }
                     connection.Close();
                 }
@@ -129,14 +127,14 @@ namespace Expense_Tracker_v2
             }
             else
             {
-                if(MessageBox.Show("Are you sure you want to Update ID: " + getID + "?", "Confirmation Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
+                if (MessageBox.Show("Are you sure you want to Update ID: " + getID + "?", "Confirmation Message", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes)
                 {
                     using (SqlConnection connection = new SqlConnection(stringConnection))
                     {
                         connection.Open();
 
                         string updateData = "UPDATE expenses SET category = @cat, item = @item" +
-                            ", expense = @expense, description = @desc, date_expense = @date_ex WHERE id = @id";
+                            ", expense = @expense, description = @desc, date_expense = @date_ex WHERE id = @id AND UserId = @userId";
 
                         using (SqlCommand cmd = new SqlCommand(updateData, connection))
                         {
@@ -146,17 +144,16 @@ namespace Expense_Tracker_v2
                             cmd.Parameters.AddWithValue("@desc", expense_description.Text);
                             cmd.Parameters.AddWithValue("@date_ex", expense_date.Value);
                             cmd.Parameters.AddWithValue("@id", getID);
+                            cmd.Parameters.AddWithValue("@userId", Loginn.UserId);
 
                             cmd.ExecuteNonQuery();
                             clearFields();
 
                             MessageBox.Show("Expense updated successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
-
                         }
                         connection.Close();
                     }
-                }  
+                }
             }
             displayExpenseData();
         }
@@ -192,18 +189,17 @@ namespace Expense_Tracker_v2
                     {
                         connection.Open();
 
-                        string deleteData = "DELETE FROM expenses WHERE id = @id";
+                        string deleteData = "DELETE FROM expenses WHERE id = @id AND UserId = @userId";
 
                         using (SqlCommand cmd = new SqlCommand(deleteData, connection))
                         {
                             cmd.Parameters.AddWithValue("@id", getID);
+                            cmd.Parameters.AddWithValue("@userId", Loginn.UserId);
 
                             cmd.ExecuteNonQuery();
                             clearFields();
 
                             MessageBox.Show("Expense deleted successfully", "Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
-
-
                         }
                         connection.Close();
                     }
